+++ packages/js/spec/interpreters.spec.ts
@@ -21,7 +21,7 @@ import {
   nor,
   not,
   createJsInterpreter,
-  compare as defaultCompare
+  compare as defaultCompare,
 } from '../src'
 import {
   includeExamplesForFieldCondition,
@@ -245,6 +245,13 @@ describe('Condition Interpreter', () => {
       expect(interpret(condition, { items: [{ age: 1 }] })).to.be.true
       expect(interpret(condition, { items: [{ name: 'test' }, { age: 1 }] })).to.be.true
     })
+
+    it('can check that element item exists', () => {
+      const condition = new Field('exists', 'items.0', false)
+
+      expect(interpret(condition, { items: [] })).to.be.true
+      expect(interpret(condition, { items: [1] })).to.be.false
+    })
   })
 
   describe('mod', () => {
@@ -314,6 +321,7 @@ describe('Condition Interpreter', () => {
       const condition = new Field('size', 'items.a', 2)
 
       expect(interpret(condition, { items: [{ a: [2, 3] }, { a: [] }, {}] })).to.be.true
+      expect(interpret(condition, { items: [{ a: [2, 3] }, { a: [4] }, { a: [2] }] })).to.be.true
       expect(interpret(condition, { items: [5, 4] })).to.be.false
       expect(interpret(condition, { items: { a: [2, 3] } })).to.be.true
     })
@@ -344,6 +352,15 @@ describe('Condition Interpreter', () => {
 
       expect(interpret(condition, item('ucast@github.com'))).to.be.true
       expect(interpret(condition, item('some text'))).to.be.false
+    })
+
+    it('return false if key is `undefined` or `null`', () => {
+      const condition = new Field('regex', 'email', /i/)
+
+      expect(interpret(condition, { email: 'ucast@github.com', })).to.be.true
+      expect(interpret(condition, { email: undefined, })).to.be.false
+      expect(interpret(condition, { email: null, })).to.be.false
+      expect(interpret(condition, {})).to.be.false
     })
   })
 
@@ -416,6 +433,15 @@ describe('Condition Interpreter', () => {
 
       expect(compare).to.have.been.called.with(1, 1)
     })
+
+    it('checks that at least one item from array satisfies condition', () => {
+      const condition = new Field('within', 'items.specs.age', [1])
+
+      expect(interpret(condition, { items: [] })).to.be.false
+      expect(interpret(condition, { items: [{ specs: [{ age: 2 }] }] })).to.be.false
+      expect(interpret(condition, { items: [{ specs: [{ age: 1 }] }] })).to.be.true
+      expect(interpret(condition, { items: [{ specs: 'test' }, { specs: [{ age: 1 }] }] })).to.be.true
+    })
   })
 
   describe('nin', () => {
@@ -467,6 +493,16 @@ describe('Condition Interpreter', () => {
 
       expect(compare).to.have.been.called.with(1, 1)
     })
+
+    it('checks that at least one item from array satisfies condition', () => {
+      const condition = new Field('all', 'items.prices', [1, 2, 3])
+
+      expect(interpret(condition, { items: [] })).to.be.false
+      expect(interpret(condition, { items: [{ prices: [1, 2] }] })).to.be.false
+      expect(interpret(condition, { items: [{ prices: condition.value }] })).to.be.true
+      expect(interpret(condition, { items: [{ names: ['test'] }, { prices: condition.value }] }))
+        .to.be.true
+    })
   })
 
   describe('elemMatch', () => {
@@ -480,15 +516,30 @@ describe('Condition Interpreter', () => {
     })
 
     it('checks that field value item matches all conditions', () => {
-      const condition = new Field('elemMatch', 'items', new CompoundCondition('and', [
-        new Field('eq', 'age', 1),
-        new Field('eq', 'active', true)
-      ]))
+      const condition = new Field(
+        'elemMatch',
+        'items',
+        new CompoundCondition('and', [new Field('eq', 'age', 1), new Field('eq', 'active', true)])
+      )
       const item = (items: object[]) => ({ items })
 
       expect(interpret(condition, item([{ age: 1 }]))).to.be.false
       expect(interpret(condition, item([]))).to.be.false
       expect(interpret(condition, item([{ age: 1 }, { active: true }]))).to.be.false
+      expect(interpret(condition, item([{ age: 1 }, { age: 1, active: true }]))).to.be.true
+    })
+
+    it('checks that at least one item from array satisfies condition', () => {
+      const condition = new Field(
+        'elemMatch',
+        'items.subItems',
+        new CompoundCondition('and', [new Field('eq', 'age', 1), new Field('eq', 'active', true)])
+      )
+      const item = (subItems: object[]) => ({ items: [{ subItems }] })
+
+      expect(interpret(condition, item([{ age: 1 }]))).to.be.false
+      expect(interpret(condition, item([]))).to.be.false
+      expect(interpret(condition, item([{ age: 1, active: true }]))).to.be.true
       expect(interpret(condition, item([{ age: 1 }, { age: 1, active: true }]))).to.be.true
     })
   })
@@ -497,9 +548,7 @@ describe('Condition Interpreter', () => {
     const interpret = createJsInterpreter({ not, eq })
 
     it('inverts nested condition', () => {
-      const condition = new CompoundCondition('not', [
-        new Field('eq', 'age', 12),
-      ])
+      const condition = new CompoundCondition('not', [new Field('eq', 'age', 12)])
 
       expect(interpret(condition, { age: 12 })).to.be.false
       expect(interpret(condition, { age: 13 })).to.be.true
@@ -512,7 +561,7 @@ describe('Condition Interpreter', () => {
     it('combines conditions using logical "and"', () => {
       const condition = new CompoundCondition('and', [
         new Field('eq', 'age', 1),
-        new Field('eq', 'active', true)
+        new Field('eq', 'active', true),
       ])
 
       expect(interpret(condition, { age: 1, active: true })).to.be.true
@@ -526,7 +575,7 @@ describe('Condition Interpreter', () => {
     it('combines conditions using logical "or"', () => {
       const condition = new CompoundCondition('or', [
         new Field('eq', 'age', 1),
-        new Field('eq', 'active', true)
+        new Field('eq', 'active', true),
       ])
 
       expect(interpret(condition, { age: 1, active: true })).to.be.true
@@ -542,7 +591,7 @@ describe('Condition Interpreter', () => {
     it('combines conditions using logical "not or"', () => {
       const condition = new CompoundCondition('nor', [
         new Field('eq', 'age', 1),
-        new Field('eq', 'active', true)
+        new Field('eq', 'active', true),
       ])
 
       expect(interpret(condition, { age: 1, active: true })).to.be.false