+++ packages/js/src/utils.ts
@@ -4,9 +4,13 @@ import { JsInterpretationOptions, JsInterpreter } from './types';
 export type AnyObject = Record<PropertyKey, unknown>;
 export type GetField = (object: any, field: string) => any;
 
-export function includes<T>(items: T[], value: T, equal: JsInterpretationOptions['equal']): boolean {
+export function includes<T>(
+  items: T[],
+  value: T,
+  compare: JsInterpretationOptions['compare']
+): boolean {
   for (let i = 0, length = items.length; i < length; i++) {
-    if (equal(items[i], value)) {
+    if (compare(items[i], value) === 0) {
       return true;
     }
   }
@@ -14,13 +18,25 @@ export function includes<T>(items: T[], value: T, equal: JsInterpretationOptions
   return false;
 }
 
+export function isArrayAndNotNumericField<T>(object: T | T[], field: string): object is T[] {
+  return Array.isArray(object) && Number.isNaN(Number(field));
+}
+
 function getField<T extends AnyObject>(object: T | T[], field: string, get: GetField) {
-  if (Array.isArray(object)) {
-    const items = object.map(item => get(item, field));
-    return Object.defineProperty(items, 'projected', { value: true });
+  if (!isArrayAndNotNumericField(object, field)) {
+    return get(object, field);
   }
 
-  return get(object, field);
+  let result: unknown[] = [];
+
+  for (let i = 0; i < object.length; i++) {
+    const value = get(object[i], field);
+    if (typeof value !== 'undefined') {
+      result = result.concat(value);
+    }
+  }
+
+  return result;
 }
 
 export function getValueByPath(object: AnyObject, field: string, get: GetField) {