+++ packages/objection/src/index.ts
@@ -1,3 +1,0 @@
-export * from './interpreters';
-export * from './interpreter';
-export * from './defaults';