+++ rollup.config.js
@@ -2,17 +2,31 @@ import babel from '@rollup/plugin-babel';
 import resolve from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
 import { terser } from 'rollup-plugin-terser';
-import { dirname } from 'path';
+import { dirname, basename, extname } from 'path';
+
+const output = (config) => {
+  let prop = 'dir';
+  let path = `dist/${config.id}${config.subpath}`;
+
+  if (config.ext) {
+    prop = 'file';
+    path += `/${basename(config.input, extname(config.input))}${config.ext}`;
+  }
+
+  return {
+    [prop]: path,
+    format: config.format,
+    name: config.name,
+    globals: config.globals,
+  };
+};
 
 const build = config => ({
-  input: config.input || 'src/index.ts',
+  input: config.input,
   external: config.external,
   output: {
     sourcemap: true,
-    format: config.format,
-    dir: `dist/${config.id}${config.subpath}`,
-    name: config.name,
-    globals: config.globals,
+    ...output(config),
     plugins: [
       config.minify
         ? terser({
@@ -43,6 +57,7 @@ const build = config => ({
 function parseOptions(overrideOptions) {
   const options = {
     external: [],
+    input: 'src/index.ts',
     subpath: '',
     useInputSourceMaps: !!process.env.USE_SRC_MAPS,
     minify: process.env.NODE_ENV === 'production'
@@ -77,7 +92,7 @@ function parseOptions(overrideOptions) {
 
 export default (overrideOptions) => {
   const builds = [
-    { id: 'es6m', type: 'es6', format: 'es' },
+    { id: 'es6m', type: 'es6', format: 'es', ext: '.mjs' },
     { id: 'es6c', type: 'es6', format: 'cjs' },
     { id: 'es5m', type: 'es5', format: 'es' },
     { id: 'umd', type: 'es5', format: 'umd', name: overrideOptions.name },